refactor(tile): fork the runtime coefficient carrier out of Projection
Nine of `Projection`'s methods answer one question the other thirty-six do not:
which of its scales, offsets and divisors are only known at runtime, and where
each one sits in the flat carriers the launch packs them into. The launch fills
those carriers by walking the maps in order and the kernel indexes them with
the same functions, so the ordering rule is the whole content, and it was
stated nowhere.

It is now `physical/projection/carrier.rs`, with the rule in its module doc.
As with `space/share.rs`, the move made the code go through the accessors
rather than the private `physical` field, which wanted one new iterator
(`axis_maps`) and lost three slice-index expressions.

No behaviour change.
